+++ plugin/src/web/server/api/system/homedir.get.ts
@@ -1,7 +1,0 @@
-import { homedir } from 'os'
-
-export default defineEventHandler(() => {
-  return {
-    homedir: homedir()
-  }
-})